Add a single implementation for GetBdxStatusCodeFromChipError
