+++ src/protocols/bdx/BdxMessages.h
@@ -28,6 +28,7 @@
 #include <lib/support/BufferWriter.h>
 #include <lib/support/CodeUtils.h>
 #include <protocols/Protocols.h>
+#include <protocols/bdx/StatusCode.h>
 #include <system/SystemPacketBuffer.h>
 
 namespace chip {
@@ -49,24 +50,6 @@ enum class MessageType : uint8_t
     BlockAck           = 0x13,
     BlockAckEOF        = 0x14,
     BlockQueryWithSkip = 0x15,
-};
-
-enum class StatusCode : uint16_t
-{
-    kLengthTooLarge             = 0x0012,
-    kLengthTooShort             = 0x0013,
-    kLengthMismatch             = 0x0014,
-    kLengthRequired             = 0x0015,
-    kBadMessageContents         = 0x0016,
-    kBadBlockCounter            = 0x0017,
-    kUnexpectedMessage          = 0x0018,
-    kResponderBusy              = 0x0019,
-    kTransferFailedUnknownError = 0x001F,
-    kTransferMethodNotSupported = 0x0050,
-    kFileDesignatorUnknown      = 0x0051,
-    kStartOffsetNotSupported    = 0x0052,
-    kVersionNotSupported        = 0x0053,
-    kUnknown                    = 0x005F,
 };
 
 enum class TransferControlFlags : uint8_t
